Make get_jvm_index return the index by int

+++ src/semantic_analyzer/symbol_table/symbol.cpp
@@ -175,9 +175,9 @@ codesh::ast::local_variable_declaration_ast_node *codesh::semantic_analyzer::loc
     return producing_node;
 }
 
-size_t codesh::semantic_analyzer::local_variable_symbol::get_jvm_index() const
+int codesh::semantic_analyzer::local_variable_symbol::get_jvm_index() const
 {
-    return index;
+    return static_cast<int>(index);
 }
 
 codesh::semantic_analyzer::method_overloads_symbol::method_overloads_symbol(
